# Optimize `getLedgers` queries by only decoding the header