--- a/cmd/stellar-rpc/internal/db/ledger_test.go
+++ b/cmd/stellar-rpc/internal/db/ledger_test.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"context"
+	"io"
 	"path"
 	"testing"
 
@@ -143,7 +144,7 @@ func TestGetLedgerRange_NonEmptyDB(t *testing.T) {
 
 func TestGetLedgerRange_SingleDBRow(t *testing.T) {
 	db := NewTestDB(t)
-	ctx := context.TODO()
+	ctx := t.Context()
 
 	writer := NewReadWriter(logger, db, interfaces.MakeNoOpDeamon(), 10, 10, passphrase)
 	write, err := writer.NewTx(ctx)
@@ -186,8 +187,7 @@ func BenchmarkGetLedgerRange(b *testing.B) {
 	testDB, lcms := setupBenchmarkingDB(b)
 	reader := NewLedgerReader(testDB)
 
-	b.ResetTimer()
-	for range b.N {
+	for b.Loop() {
 		ledgerRange, err := reader.GetLedgerRange(context.TODO())
 		require.NoError(b, err)
 		assert.Equal(b, lcms[0].LedgerSequence(), ledgerRange.FirstLedger.Sequence)
@@ -204,12 +204,15 @@ func BenchmarkBatchGetLedgers(b *testing.B) {
 
 	start := uint32(1334)
 	end := start + uint32(batchSize) - 1
-	b.ResetTimer()
-	for range b.N {
+
+	for b.Loop() {
 		ledgers, err := readTx.BatchGetLedgers(b.Context(), start, end)
 		require.NoError(b, err)
-		assert.Equal(b, lcms[0].LedgerSequence(), ledgers[0].LedgerSequence())
-		assert.Equal(b, lcms[batchSize-1].LedgerSequence(), ledgers[batchSize-1].LedgerSequence())
+
+		hdrFirst := ledgers[0].Header.Header
+		hdrLast := ledgers[batchSize-1].Header.Header
+		assert.EqualValues(b, lcms[0].LedgerSequence(), hdrFirst.LedgerSeq)
+		assert.EqualValues(b, lcms[batchSize-1].LedgerSequence(), hdrLast.LedgerSeq)
 	}
 }
 
@@ -227,9 +230,11 @@ func NewTestDB(tb testing.TB) *DB {
 func setupBenchmarkingDB(b *testing.B) (*DB, []xdr.LedgerCloseMeta) {
 	testDB := NewTestDB(b)
 	logger := log.DefaultLogger
+	logger.SetOutput(io.Discard)
+
 	writer := NewReadWriter(logger, testDB, interfaces.MakeNoOpDeamon(),
 		100, 1_000_000, passphrase)
-	write, err := writer.NewTx(context.TODO())
+	write, err := writer.NewTx(b.Context())
 	require.NoError(b, err)
 
 	lcms := make([]xdr.LedgerCloseMeta, 0, 100_000)