--- a/cmd/stellar-rpc/internal/methods/get_ledgers.go
+++ b/cmd/stellar-rpc/internal/methods/get_ledgers.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/creachadair/jrpc2"
 	"github.com/stellar/go/support/log"
-	"github.com/stellar/go/xdr"
 
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db"
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/rpcdatastore"
@@ -37,8 +36,10 @@ func NewGetLedgersHandler(ledgerReader db.LedgerReader, maxLimit, defaultLimit u
 	}).getLedgers)
 }
 
-// getLedgers fetch ledgers and relevant metadata from DB and falling back to the remote rpcdatastore if necessary.
-func (h ledgersHandler) getLedgers(ctx context.Context, request protocol.GetLedgersRequest,
+// getLedgers fetch ledgers and relevant metadata from DB and falling back to
+// the remote rpcdatastore if necessary.
+func (h ledgersHandler) getLedgers(
+	ctx context.Context, request protocol.GetLedgersRequest,
 ) (protocol.GetLedgersResponse, error) {
 	readTx, err := h.ledgerReader.NewTx(ctx)
 	if err != nil {
@@ -159,99 +160,113 @@ func (h ledgersHandler) parseCursor(cursor string, ledgerRange protocol.LedgerSe
 //  1. Entire range is available in local db.
 //  2. Entire range is unavailable in local db so fetch fully from datastore.
 //  3. Range partially available in the local db with the rest fetched from the datastore.
-func (h ledgersHandler) fetchLedgers(ctx context.Context, start uint32,
-	end uint32, format string, readTx db.LedgerReaderTx, localLedgerRange protocol.LedgerSeqRange,
+func (h ledgersHandler) fetchLedgers(
+	ctx context.Context,
+	start, end uint32, format string,
+	readTx db.LedgerReaderTx,
+	localLedgerRange protocol.LedgerSeqRange,
 ) ([]protocol.LedgerInfo, error) {
-	fetchFromLocalDB := func(start uint32, end uint32) ([]xdr.LedgerCloseMeta, error) {
+	limit := end - start + 1
+	result := make([]protocol.LedgerInfo, 0, limit)
+
+	addToResult := func(ledgers []db.LedgerMetadataChunk) error {
+		for _, chunk := range ledgers {
+			if len(result) >= int(limit) {
+				break
+			}
+
+			if info, err := parseLedgerInfo(chunk, format); err != nil {
+				return &jrpc2.Error{
+					Code: jrpc2.InternalError,
+					Message: fmt.Sprintf("error processing ledger %d: %v",
+						chunk.Header.Header.LedgerSeq, err),
+				}
+			} else {
+				result = append(result, info)
+			}
+		}
+		return nil
+	}
+
+	fetchFromLocalDB := func(start, end uint32) error {
 		ledgers, err := readTx.BatchGetLedgers(ctx, start, end)
 		if err != nil {
-			return nil, &jrpc2.Error{
+			return &jrpc2.Error{
 				Code:    jrpc2.InternalError,
 				Message: fmt.Sprintf("error fetching ledgers from db: %v", err),
 			}
 		}
-		return ledgers, nil
+
+		return addToResult(ledgers)
 	}
 
-	fetchFromDatastore := func(start uint32, end uint32) ([]xdr.LedgerCloseMeta, error) {
+	fetchFromDatastore := func(start, end uint32) error {
 		if h.datastoreLedgerReader == nil {
-			return nil, &jrpc2.Error{
+			return &jrpc2.Error{
 				Code:    jrpc2.InvalidParams,
 				Message: "datastore ledger reader not configured",
 			}
 		}
 		ledgers, err := h.datastoreLedgerReader.GetLedgers(ctx, start, end)
 		if err != nil {
-			return nil, &jrpc2.Error{
+			return &jrpc2.Error{
 				Code:    jrpc2.InternalError,
 				Message: fmt.Sprintf("error fetching ledgers from datastore: %v", err),
 			}
 		}
-		return ledgers, nil
+
+		for _, ledger := range ledgers {
+			raw, err := ledger.MarshalBinary()
+			if err != nil {
+				return &jrpc2.Error{
+					Code:    jrpc2.InternalError,
+					Message: fmt.Sprintf("error fetching ledgers from datastore: %v", err),
+				}
+			}
+
+			if err := addToResult([]db.LedgerMetadataChunk{
+				{Lcm: raw, Header: ledger.LedgerHeaderHistoryEntry()},
+			}); err != nil {
+				return err
+			}
+		}
+		return nil
 	}
 
-	limit := end - start + 1
-	ledgers := make([]xdr.LedgerCloseMeta, 0, limit)
 	switch {
 	case start >= localLedgerRange.FirstLedger:
 		// entire range is available in local DB
-		localLedgers, err := fetchFromLocalDB(start, end)
-		if err != nil {
+		if err := fetchFromLocalDB(start, end); err != nil {
 			return nil, err
 		}
-		ledgers = append(ledgers, localLedgers...)
 
 	case end < localLedgerRange.FirstLedger:
 		// entire range is unavailable locally so fetch everything from datastore
-		dataStoreLedgers, err := fetchFromDatastore(start, end)
-		if err != nil {
+		if err := fetchFromDatastore(start, end); err != nil {
 			return nil, err
 		}
-		ledgers = append(ledgers, dataStoreLedgers...)
 
 	default:
 		// part of the ledger range is available locally so fetch local ledgers from db,
 		// and the rest from the datastore.
-
-		dataStoreLedgers, err := fetchFromDatastore(start, localLedgerRange.FirstLedger-1)
-		if err != nil {
+		if err := fetchFromDatastore(start, localLedgerRange.FirstLedger-1); err != nil {
 			return nil, err
 		}
-		ledgers = append(ledgers, dataStoreLedgers...)
-
-		localLedgers, err := fetchFromLocalDB(localLedgerRange.FirstLedger, end)
-		if err != nil {
+		if err := fetchFromLocalDB(localLedgerRange.FirstLedger, end); err != nil {
 			return nil, err
 		}
-		ledgers = append(ledgers, localLedgers...)
-	}
-
-	// convert raw lcm to protocol.LedgerInfo
-	result := make([]protocol.LedgerInfo, 0, limit)
-	for _, ledger := range ledgers {
-		if len(result) >= int(limit) {
-			break
-		}
-
-		ledgerInfo, err := h.parseLedgerInfo(ledger, format)
-		if err != nil {
-			return nil, &jrpc2.Error{
-				Code:    jrpc2.InternalError,
-				Message: fmt.Sprintf("error processing ledger %d: %v", ledger.LedgerSequence(), err),
-			}
-		}
-		result = append(result, ledgerInfo)
 	}
 
 	return result, nil
 }
 
 // parseLedgerInfo extracts and formats the ledger metadata and header information.
-func (h ledgersHandler) parseLedgerInfo(ledger xdr.LedgerCloseMeta, format string) (protocol.LedgerInfo, error) {
+func parseLedgerInfo(ledger db.LedgerMetadataChunk, format string) (protocol.LedgerInfo, error) {
+	header := ledger.Header
 	ledgerInfo := protocol.LedgerInfo{
-		Hash:            ledger.LedgerHash().HexString(),
-		Sequence:        ledger.LedgerSequence(),
-		LedgerCloseTime: ledger.LedgerCloseTime(),
+		Hash:            header.Hash.HexString(),
+		Sequence:        uint32(header.Header.LedgerSeq),
+		LedgerCloseTime: int64(header.Header.ScpValue.CloseTime),
 	}
 
 	// Format the data according to the requested format (JSON or XDR)
@@ -262,18 +277,14 @@ func (h ledgersHandler) parseLedgerInfo(ledger xdr.LedgerCloseMeta, format strin
 		if convErr != nil {
 			return ledgerInfo, convErr
 		}
-	default:
-		closeMetaB, err := ledger.MarshalBinary()
-		if err != nil {
-			return protocol.LedgerInfo{}, fmt.Errorf("error marshaling ledger close meta: %w", err)
-		}
 
-		headerB, err := ledger.LedgerHeaderHistoryEntry().MarshalBinary()
+	default:
+		headerB, err := ledger.Header.MarshalBinary()
 		if err != nil {
 			return protocol.LedgerInfo{}, fmt.Errorf("error marshaling ledger header: %w", err)
 		}
 
-		ledgerInfo.LedgerMetadata = base64.StdEncoding.EncodeToString(closeMetaB)
+		ledgerInfo.LedgerMetadata = base64.StdEncoding.EncodeToString(ledger.Lcm)
 		ledgerInfo.LedgerHeader = base64.StdEncoding.EncodeToString(headerB)
 	}
 	return ledgerInfo, nil