### What
Instead of decoding the raw binary data directly into an `xdr.LedgerCloseMeta`, this modifies the code to only decode the `xdr.LedgerHeaderHistoryEntry` out of the data which comes first in the layout. The rest of the meta is passed around directly as a `[]byte` until the final step, either converting it to base64 directly or to JSON.

Rudimentary benchmarks show a ~40% improvement over the baseline:

```
$ git switch main
$ go test -failfast -run=None -bench=BenchmarkBatchGetLedgers ./cmd/stellar-rpc/internal/db
goos: linux
goarch: amd64
pkg: github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db
cpu: 13th Gen Intel(R) Core(TM) i7-1360P
BenchmarkBatchGetLedgers-16    	git switch -    1788	    773248 ns/op
PASS
ok  	github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db	7.652s

$ git switch getLedgers_depthCheck
$ go test -failfast -run=None -bench=BenchmarkBatchGetLedgers ./cmd/stellar-rpc/internal/db
goos: linux
goarch: amd64
pkg: github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db
cpu: 13th Gen Intel(R) Core(TM) i7-1360P
BenchmarkBatchGetLedgers-16    	    1939	    530915 ns/op
PASS
ok  	github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db	3.325s
```

### Why
We have observed performance issues with `getLedgers` due to their density and would like to improve them.

### Known limitations
This needs to be evaluated against real load to see the improvement.